Implement Palmer-style precues
Under this method, all stimuli are shown on every trial, and setsize is
manipulated by precuing a subset of the locations.  Whether or not to
precue is controlled by a variable, palmerFlag.  Precues are presented on
all experimental runs, but colors are used to make them appear identical to
normal pedastals when palmerFlag == 0.  This manner of presenting precues
makes timing consistent across trial types.