Aborted attempt to add method of constant stimuli
The idea was to add a flag to enable presentation of noise level using
method of constant stimuli (MCS).  The implementation here is rather
hackish and incomplete, and completing it would require a lot of
complicated code for all the possible combinations of flags.

Also, there's a way to enter parameters into the dialog box to effectively
run an MCS version of the experiment: Enter a stimulus set multiple times,
then specify different noise levels for each one.

I am saving this effort to a separate branch in order to have a record of
the attempt.