Clean up handling of noise and number of trials
These are two interrelated issues because noise was included as a factor on
practice and experimental trials.  Now we treat noise as a fixed value,
with one value per stimulus set.  On all practice phases and experimental
phases that don't follow a staircase phase, the supplied noise values are
used; on staircase phases, the first supplied noise value is used; and on
experimental phases that follow a staircase phase, the final staircase
values are used.

Experimental phase length is now determined by number of replicates
requested, while the practice phase length is determined by the total
number of practice trials requested.

Also, ensure that the supplied noise levels make sense.  There should be
one value or one value per stimulus set; if not, generate an error message.