Updated to version run in Crossover04
Crossover04 was run using code that I had modified and committed here.
Evan Palmer fixed it to run the actual experiment, and I had never received
the final copy he used.  It appears to be the version that is described in
the initial Crossover submission.

This version adds an experimental block after the staircase block that runs
with the staircased noise values.  It also allows for a run without a
staircase, and modifies several default values.

The only change from Evan's version is to Unix line endings.