Switch to Kaernbach's (1990) SIAM adaptive procedure
We tested the original simple up-down staircase, but the data didn't
overlap properly.  One possible problem with this staircase is that it
adjusts the noise after a correct rejection, which can lead to biased
threshold estimates when the subject responds correctly to a series of
target-absent trials in which the target would be invisible.

The SIAM procedure may be better because it does not adjust the noise after
correct rejections.  The current version uses the same hit adjustment as
the original staircase used for all correct trials, and still targets an
accuracy of 80%.  The values may need adjusting for a more fine-tuned
staircase.

Also advance version RC number.