Filter out trials with RT >= 10000 ms.
Change from filtering out RT >= 2000 ms.  Both make sense, and this change
has little noticeable effect on the model fits.  The change was based on
boxplots and stem-and-leaf plots of RT across subjects and IVs.  These
plots indicate that the RTs over 10 s are blatant outliers, whereas those
between 2-10 s could just be the tails of a gamma distribution (e.g., EMP
has one RT near 115 s).

I should do a more careful evaluation of the effects of outliers, but a
data set this huge is unlikely to be affected much.  Nevertheless, some of
the subjects were distracted occasionally, so some outliers are real.