Allow different starting capacities.
Each subject can have a different starting capacity for each stimulus set.
The current values lead to good fits across subjects.  This is justifiable
because it's just leading to the best possible fit.

Equivalent changes to the other parameters or to the unlimited-capacity
model are unnecessary because those parameters have continuous, linear
effects on the model.  Capacity has distinctly different effects as it
crosses each setsize.