Add analyses to test effects of setsize on criterion.
One analysis for the unlimited-capacity models, and one for the
limited-capacity models.

The analysis files include some minor hacking to get around an awkward
library-loading issue.  The ANOVA run in this scipt uses ezANOVA from the
ez package.  That package loads the reshape2 package, which has the useful
melt() function.  However, ez then loads the ggplot2 package, which itself
loads the older reshape package in front of reshape2, forcing the use of
the older melt() functions.

I tried the following hacking to get around this problem, but ended up not
using it:

    if (any(search() == "package:reshape2"))
        detach("package:reshape2", force=TRUE)
    library("reshape2", quietly=TRUE, warn.conflicts=FALSE)